+++ include/proxysql_structs.h
@@ -476,7 +476,7 @@ struct _global_variables_t {
 
 	unsigned char protocol_version;
 	char *mysql_server_version;
-	uint16_t server_capabilities;
+	uint32_t server_capabilities;
 	uint8_t server_language;
 	uint16_t server_status;
 
@@ -625,7 +625,7 @@ __thread int mysql_thread___connect_timeout_server_max;
 __thread int mysql_thread___query_processor_iterations;
 __thread int mysql_thread___query_processor_regex;
 __thread int mysql_thread___reset_connection_algorithm;
-__thread uint16_t mysql_thread___server_capabilities;
+__thread uint32_t mysql_thread___server_capabilities;
 __thread uint8_t mysql_thread___default_charset;
 __thread int mysql_thread___poll_timeout;
 __thread int mysql_thread___poll_timeout_on_failure;
@@ -741,7 +741,7 @@ extern __thread int mysql_thread___connect_timeout_server_max;
 extern __thread int mysql_thread___query_processor_iterations;
 extern __thread int mysql_thread___query_processor_regex;
 extern __thread int mysql_thread___reset_connection_algorithm;
-extern __thread uint16_t mysql_thread___server_capabilities;
+extern __thread uint32_t mysql_thread___server_capabilities;
 extern __thread uint8_t mysql_thread___default_charset;
 extern __thread int mysql_thread___poll_timeout;
 extern __thread int mysql_thread___poll_timeout_on_failure;
